+++ packages/headless/src/features/facets/facet-set/facet-set-analytics-actions.ts
@@ -5,20 +5,18 @@ import {
   facetIdDefinition,
   requiredNonEmptyString,
 } from '../generic/facet-actions-validation';
-import {SearchAppState} from '../../../state/search-app-state';
-import {getFacetSetInitialState} from './facet-set-state';
-import {getDateFacetSetInitialState} from '../range-facets/date-facet-set/date-facet-set-state';
-import {getNumericFacetSetInitialState} from '../range-facets/numeric-facet-set/numeric-facet-set-state';
-import {getCategoryFacetSetInitialState} from '../category-facet-set/category-facet-set-state';
 import {Value} from '@coveo/bueno';
 import {
   AnalyticsType,
   makeAnalyticsAction,
 } from '../../analytics/analytics-utils';
 import {
+  buildFacetBaseMetadata,
   FacetSelectionChangeMetadata,
   FacetUpdateSortMetadata,
-  SectionNeededForFacetMetadata,
+  buildFacetStateMetadata,
+  getStateNeededForFacetMetadata,
+  buildFacetSelectionChangeMetadata,
 } from './facet-set-analytics-actions-utils';
 
 /**
@@ -67,11 +65,10 @@ export const logFacetSearch = (facetId: string) =>
     AnalyticsType.Search,
     (client, state) => {
       validatePayload(facetId, facetIdDefinition);
-      const metadata = buildFacetBaseMetadata(
-        facetId,
-        getStateNeededForFacetMetadata(state)
-      );
-      return client.logFacetSearch(metadata);
+      const stateForAnalytics = getStateNeededForFacetMetadata(state);
+      const metadata = buildFacetBaseMetadata(facetId, stateForAnalytics);
+      const facetState = buildFacetStateMetadata(stateForAnalytics);
+      return client.logFacetSearch(metadata, facetState);
     }
   )();
 /**
@@ -89,15 +86,15 @@ export const logFacetUpdateSort = (payload: FacetUpdateSortMetadata) =>
           required: true,
         }),
       });
+
       const {facetId, criterion} = payload;
+      const stateForAnalytics = getStateNeededForFacetMetadata(state);
 
-      const base = buildFacetBaseMetadata(
-        facetId,
-        getStateNeededForFacetMetadata(state)
-      );
+      const base = buildFacetBaseMetadata(facetId, stateForAnalytics);
       const metadata = {...base, criteria: criterion};
+      const facetState = buildFacetStateMetadata(stateForAnalytics);
 
-      return client.logFacetUpdateSort(metadata);
+      return client.logFacetUpdateSort(metadata, facetState);
     }
   )();
 
@@ -111,11 +108,12 @@ export const logFacetClearAll = (facetId: string) =>
     AnalyticsType.Search,
     (client, state) => {
       validatePayload(facetId, facetIdDefinition);
-      const metadata = buildFacetBaseMetadata(
-        facetId,
-        getStateNeededForFacetMetadata(state)
-      );
-      return client.logFacetClearAll(metadata);
+
+      const stateForAnalytics = getStateNeededForFacetMetadata(state);
+      const metadata = buildFacetBaseMetadata(facetId, stateForAnalytics);
+      const facetState = buildFacetStateMetadata(stateForAnalytics);
+
+      return client.logFacetClearAll(metadata, facetState);
     }
   )();
 
@@ -133,12 +131,14 @@ export const logFacetSelect = (payload: FacetSelectionChangeMetadata) =>
         facetValue: requiredNonEmptyString,
       });
 
+      const stateForAnalytics = getStateNeededForFacetMetadata(state);
       const metadata = buildFacetSelectionChangeMetadata(
         payload,
-        getStateNeededForFacetMetadata(state)
+        stateForAnalytics
       );
+      const facetState = buildFacetStateMetadata(stateForAnalytics);
 
-      return client.logFacetSelect(metadata);
+      return client.logFacetSelect(metadata, facetState);
     }
   )();
 
@@ -155,11 +155,14 @@ export const logFacetDeselect = (payload: FacetSelectionChangeMetadata) =>
         facetId: facetIdDefinition,
         facetValue: requiredNonEmptyString,
       });
+      const stateForAnalytics = getStateNeededForFacetMetadata(state);
       const metadata = buildFacetSelectionChangeMetadata(
         payload,
-        getStateNeededForFacetMetadata(state)
+        stateForAnalytics
       );
-      return client.logFacetDeselect(metadata);
+      const facetState = buildFacetStateMetadata(stateForAnalytics);
+
+      return client.logFacetDeselect(metadata, facetState);
     }
   )();
 
@@ -183,40 +186,3 @@ export const logFacetBreadcrumb = (payload: FacetSelectionChangeMetadata) =>
       return client.logBreadcrumbFacet(metadata);
     }
   )();
-
-function buildFacetSelectionChangeMetadata(
-  payload: FacetSelectionChangeMetadata,
-  state: SectionNeededForFacetMetadata
-) {
-  const {facetId, facetValue} = payload;
-  const base = buildFacetBaseMetadata(facetId, state);
-
-  return {...base, facetValue};
-}
-
-function getStateNeededForFacetMetadata(
-  s: Partial<SearchAppState>
-): SectionNeededForFacetMetadata {
-  return {
-    facetSet: s.facetSet || getFacetSetInitialState(),
-    categoryFacetSet: s.categoryFacetSet || getCategoryFacetSetInitialState(),
-    dateFacetSet: s.dateFacetSet || getDateFacetSetInitialState(),
-    numericFacetSet: s.numericFacetSet || getNumericFacetSetInitialState(),
-  };
-}
-
-function buildFacetBaseMetadata(
-  facetId: string,
-  state: SectionNeededForFacetMetadata
-) {
-  const facet =
-    state.facetSet[facetId] ||
-    state.categoryFacetSet[facetId] ||
-    state.dateFacetSet[facetId] ||
-    state.numericFacetSet[facetId];
-
-  const facetField = facet.field;
-  const facetTitle = `${facetField}_${facetId}`;
-
-  return {facetId, facetField, facetTitle};
-}