+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.103](https://github.com/coveo/ui-kit/compare/v0.1.0-alpha.102...v0.1.0-alpha.103) (2020-12-15)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
+# [0.1.0-alpha.102](https://github.com/coveo/ui-kit/compare/v0.1.0-alpha.101...v0.1.0-alpha.102) (2020-12-15)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
+# [0.1.0-alpha.101](https://github.com/coveo/ui-kit/compare/v0.1.0-alpha.100...v0.1.0-alpha.101) (2020-12-15)
+
+
+### Features
+
+* **headless:** dispatch analyticsAction before retrying query with correction ([079fbdd](https://github.com/coveo/ui-kit/commit/079fbdd32af45cf47b5df3550567d7da267d4cba))
+
+
+
+
+
+# [0.1.0-alpha.100](https://github.com/coveo/ui-kit/compare/v0.1.0-alpha.99...v0.1.0-alpha.100) (2020-12-15)
+
+
+### Bug Fixes
+
+* **headless:** fix facet state analytics ([#346](https://github.com/coveo/ui-kit/issues/346)) ([592233f](https://github.com/coveo/ui-kit/commit/592233f8521f0798b6ccfb590bef38291e561495))
+
+
+
+
+
 # [0.1.0-alpha.99](https://github.com/coveo/ui-kit/compare/v0.1.0-alpha.98...v0.1.0-alpha.99) (2020-12-14)